Revert previous tries
