+++ DREAM Lab Frontend/package.json
@@ -29,6 +29,7 @@
     "dayjs": "^1.11.10",
     "embla-carousel-react": "^8.0.0",
     "framer-motion": "^11.0.12",
+    "gsap": "^3.12.5",
     "react": "^18.2.0",
     "react-bootstrap": "^2.10.1",
     "react-dom": "^18.2.0",
