+++ DREAM Lab Frontend/tailwind.config.js
@@ -1,7 +1,5 @@
 import { nextui } from "@nextui-org/react";
 
-import { default as flattenColorPalette } from "tailwindcss/lib/util/flattenColorPalette";
-
 /** @type {import('tailwindcss').Config} */
 export default {
     content: [
@@ -10,30 +8,8 @@ export default {
         "./node_modules/@nextui-org/theme/dist/**/*.{js,ts,jsx,tsx}",
     ],
     theme: {
-        extend: {
-            animation: {
-                scroll: "scroll var(--animation-duration, 40s) var(--animation-direction, forwards) linear infinite",
-            },
-            keyframes: {
-                scroll: {
-                    to: {
-                        transform: "translate(calc(-50% - 0.5rem))",
-                    },
-                },
-            },
-        },
+        extend: {},
     },
     darkMode: "class",
-    plugins: [nextui(), addVariablesForColors],
+    plugins: [nextui()],
 };
-
-function addVariablesForColors({ addBase, theme }) {
-    let allColors = flattenColorPalette(theme("colors"));
-    let newVars = Object.fromEntries(
-        Object.entries(allColors).map(([key, val]) => [`--${key}`, val])
-    );
-
-    addBase({
-        ":root": newVars,
-    });
-}
